other_tools/pishing_module/mail_sender/requirements.txt
python-dotenv
openpyxl
mjml
requests
pydantic
flask
fastapi
uvicorn
python-multipart
css-inline